backend/requirements.txt
fastapi[standard]
uvicorn
pillow
requests
python-dotenv
python-multipart
tweepy
